% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efm.R
\name{efm}
\alias{efm}
\title{Perform exponential factor model estimation.}
\usage{
efm(
  x,
  factor_family,
  rank,
  weights = 1,
  algo = "lapl",
  start = NULL,
  dispersion = 1,
  adam_control = adam.control(max_epoch = 5, batch_size = 32, step_size = 0.1, rho = 0,
    abs_tol = 1e-06, beta1 = 0.9, beta2 = 0.999, epislon = 10^-8),
  sample_control = sample.control(sample_size = 50, eval_size = 50),
  eval_likeli = FALSE,
  identify_ = FALSE
)
}
\arguments{
\item{x}{Data to conduct factor inference}

\item{factor_family}{Family object to specify factor loss (see \code{family})}

\item{rank}{Rank of factor model.}

\item{weights}{Entry-wise weight.}

\item{algo}{Optimization algorithm to be chosen from 1). sml (simulated maximum likelihood) 2). ps (posterior sampling) 3). lapl (laplacian approxmation).}

\item{start}{matrix of initial projection matrix \code{V}}

\item{dispersion}{dispersion parameter, effective for dispersed family, should be 1 for non-dispersed family.}

\item{adam_control}{Adam optimization control parameters, (see \code{adam.control}).}

\item{sample_control}{EFM sampling control parameters, (see \code{sample.control})}

\item{eval_likeli}{boolen, set to false to skip marginal likelihood evaluation}

\item{identify_}{boolen, set to true to enable identifiability at every iteration}
}
\value{
Estimation of projection matrix \code{V}.
}
\description{
Perform exponential factor model estimation.
}
