% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efm.R
\name{efm_identify}
\alias{efm_identify}
\title{Identify projection matrix \code{V} accrodng to \verb{DV^\\top}}
\usage{
efm_identify(V)
}
\arguments{
\item{V, }{Projection matrix to be identified}
}
\description{
Identify projection matrix \code{V} accrodng to \verb{DV^\\top}
}
