% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efm.R
\name{generate_data}
\alias{generate_data}
\title{Generate data according to exponential family distribution given parameter \verb{\\eta}}
\usage{
generate_data(family, eta, weights = 1, dispersion = 1)
}
\arguments{
\item{family}{Family object to specify factor loss.}

\item{eta}{\code{eta} to simulate data}

\item{weights}{Entry-wise weight.}

\item{dispersion}{dispersion parameter, effective for dispersed family, should be 1 for non-dispersed family.}
}
\value{
simulated data from expoential family with mean \verb{\\mu = linkinv(\\eta)}
}
\description{
Generate data according to exponential family distribution given parameter \verb{\\eta}
}
