% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efm.R
\name{sample.control}
\alias{sample.control}
\title{EFM sampling parameters}
\usage{
sample.control(sample_size = 50, eval_size = 50)
}
\arguments{
\item{sample_size}{number of Monte Carlo samples to estimate the gradients}

\item{eval_size}{number of Monte Carlo samples to evaluate the marginal likelihood}
}
\value{
list of efm sampling parameters
}
\description{
EFM sampling parameters
}
