% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efm.R
\name{SML_neglikeli}
\alias{SML_neglikeli}
\title{Using Monte Carlo to evaluate the marginalized exponential factor likelihood.}
\usage{
SML_neglikeli(
  Vt,
  factor_family,
  X,
  sample_size,
  dispersion = 1,
  weights = 1,
  print_likeli = TRUE,
  log_ = TRUE
)
}
\arguments{
\item{Vt, }{Projection matrix for evaluation.}

\item{factor_family}{Family object to specify factor loss.}

\item{X}{Data to evaluate the likelihood.}

\item{sample_size}{number of Monte Carlo samples to estimate the gradients.}

\item{dispersion}{dispersion parameter, effective for dispersed family, should be 1 for non-dispersed family.}

\item{weights}{Entry-wise weight.}

\item{print_likeli}{boolen, to print out the simulated likelihood result}

\item{log_}{boolen, to log the probability}
}
\description{
Using Monte Carlo to evaluate the marginalized exponential factor likelihood.
}
